[maglev] Cleanup TryBuildNewConsString
* Use the same heuristic as TS: Not checking for empty strings when
  either right is a cons string or left is a cons string and right not
  empty. Some graph size reduction by not having the branches.
* constant fold empty strings.
* cache length estimate for StringConcat
* check string_length_protector to avoid deopt loops.

Drive-By: Add a dcheck we don't create instructions with mismatching
          number of args.

Bug: 42204525
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/8065685
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#108612}

+++ src/maglev/maglev-graph-builder.h
@@ -1557,7 +1557,7 @@ class MaglevGraphBuilder {
   MaybeReduceResult TryBuildStringConcat(ValueNode* left, ValueNode* right);
   ReduceResult BuildStringConcat(ValueNode* left, ValueNode* right);
   ReduceResult BuildNewConsStringMap(ValueNode* left, ValueNode* right);
-  size_t StringLengthStaticLowerBound(ValueNode* string, int max_depth = 2);
+  uint32_t StringLengthStaticLowerBound(ValueNode* string, int max_depth = 2);
   MaybeReduceResult TryBuildNewConsString(
       ValueNode* left, ValueNode* right,
       AllocationType allocation_type = AllocationType::kYoung);
